refactor(fix): extract integration tests to fix/integration_tests.rs; use AVAX/USD
- Convert fix.rs → fix/mod.rs + fix/integration_tests.rs, matching the
  pattern used by the KDB and etcd adapters
- Switch instrument from EUR/USD (4001) to Avalanche/USD (100946), a
  24/7 crypto market, so the test passes outside FX trading hours
